Create a container for recording indicator
