+++ bigbluebutton-html5/imports/ui/components/nav-bar/container.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Meteor } from 'meteor/meteor';
 import { withTracker } from 'meteor/react-meteor-data';
 import { Session } from 'meteor/session';
-import Meetings, { RecordMeetings } from '/imports/api/meetings';
+import Meetings from '/imports/api/meetings';
 import Users from '/imports/api/users';
 import Auth from '/imports/ui/services/auth';
 import { meetingIsBreakout } from '/imports/ui/components/app/service';
@@ -10,10 +10,9 @@ import getFromUserSettings from '/imports/ui/services/users-settings';
 import userListService from '../user-list/service';
 import Service from './service';
 import NavBar from './component';
-import mapUser from '../../services/user/mapUser';
 
 const PUBLIC_CONFIG = Meteor.settings.public;
-
+const ROLE_MODERATOR = Meteor.settings.public.user.role_moderator;
 const NavBarContainer = ({ children, ...props }) => (
   <NavBar {...props}>
     {children}
@@ -28,7 +27,6 @@ export default withTracker(() => {
   const meetingObject = Meetings.findOne({
     meetingId,
   });
-  const recordObeject = RecordMeetings.findOne({ meetingId });
 
   if (meetingObject != null) {
     meetingTitle = meetingObject.meetingProp.name;
@@ -43,30 +41,14 @@ export default withTracker(() => {
     return hasUnreadMessages;
   };
 
-  RecordMeetings.find({ meetingId: Auth.meetingID }, { fields: { recording: 1 } }).observeChanges({
-    changed: (id, fields) => {
-      if (fields && fields.recording) {
-        this.window.parent.postMessage({ response: 'recordingStarted' }, '*');
-      }
-
-      if (fields && !fields.recording) {
-        this.window.parent.postMessage({ response: 'recordingStopped' }, '*');
-      }
-    },
-  });
-
   const currentUserId = Auth.userID;
   const { connectRecordingObserver, processOutsideToggleRecording } = Service;
   const currentUser = Users.findOne({ userId: Auth.userID });
   const openPanel = Session.get('openPanel');
   const isExpanded = openPanel !== '';
-  const amIModerator = mapUser(currentUser).isModerator;
+  const amIModerator = currentUser.role === ROLE_MODERATOR;
   const isBreakoutRoom = meetingIsBreakout();
   const hasUnreadMessages = checkUnreadMessages();
-
-  const toggleUserList = () => {
-    Session.set('isUserListOpen', !isExpanded);
-  };
 
   return {
     amIModerator,
@@ -78,11 +60,5 @@ export default withTracker(() => {
     presentationTitle: meetingTitle,
     hasUnreadMessages,
     isBreakoutRoom,
-    toggleUserList,
-    allowStartStopRecording: !!(recordObeject && recordObeject.allowStartStopRecording),
-    autoStartRecording: recordObeject && recordObeject.autoStartRecording,
-    record: recordObeject && recordObeject.record,
-    recording: recordObeject && recordObeject.recording,
-    time: recordObeject && recordObeject.time,
   };
 })(NavBarContainer);
