+++ bigbluebutton-html5/imports/ui/components/nav-bar/recording-indicator/styles.scss
@@ -155,3 +155,9 @@
     outline-style: solid;
   }
 }
+
+.presentationTitleSeparator {
+  color: var(--color-gray);
+  font-size: var(--font-size-base);
+  margin: 0 1rem;
+}
