+++ bigbluebutton-html5/imports/ui/components/nav-bar/styles.scss
@@ -40,12 +40,6 @@
   }
 }
 
-.presentationTitleSeparator {
-  color: var(--color-gray);
-  font-size: var(--font-size-base);
-  margin: 0 1rem;
-}
-
 .btnWithNotificationDot {
   position: relative;
